# API_documentation/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user": {
      "get": {
        "description": "This endpoint requires a minimun security clearance of level 5.",
        "operationId": "UserController_viewUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Here are all the User Accounts available"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get User Accounts",
        "tags": [
          "Manager"
        ]
      },
      "post": {
        "description": "This endpoint requires a minimun security clearance of level 5.",
        "operationId": "UserController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Payload to create User Account",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserWithRolePermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User Account created successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "409": {
            "description": "Conflict - Resource already exist or duplicate entry"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a new user account",
        "tags": [
          "Manager"
        ]
      }
    },
    "/user/me/permissions": {
      "get": {
        "description": "This endpoint requires a minimun security clearance of level 5.",
        "operationId": "UserController_getMyPermissions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "My user account"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "My User Account",
        "tags": [
          "Manager"
        ]
      }
    },
    "/user/role_permission": {
      "post": {
        "description": "This endpoint requires a minimun security clearance of level 5.",
        "operationId": "UserController_addRolePermission",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddUserRolePermissionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role permission added to user successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "409": {
            "description": "Conflict - Resource already exist or duplicate entry"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Add Role permissions to user",
        "tags": [
          "Manager"
        ]
      }
    },
    "/user/new_reset_token": {
      "post": {
        "description": "This endpoint requires a minimun security clearance of level 5.",
        "operationId": "UserController_newResetToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Payload for new user reset token",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserEmailResetTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset done! you can now log in!"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "409": {
            "description": "Conflict - Resource already exist or duplicate entry"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Reset token for first time log in",
        "tags": [
          "Manager"
        ]
      }
    },
    "/user/token_keys": {
      "get": {
        "description": "This endpoint requires a minimun security clearance of level 5.",
        "operationId": "UserController_viewUserKeys",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserWithRolePermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Here are all the token keys available for this user"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get the token keys for this user",
        "tags": [
          "Manager"
        ]
      }
    },
    "/user/deactivate": {
      "put": {
        "description": "This endpoint requires a minimun security clearance of level 5.",
        "operationId": "UserController_deactivateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeactivateUserAccountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User account deactivated successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Deactivate the user account",
        "tags": [
          "Manager"
        ]
      }
    },
    "/user/reactivate": {
      "put": {
        "description": "This endpoint requires a minimun security clearance of level 5.",
        "operationId": "UserController_reactivateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactivateUserAccountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User account reactivated successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Reactivate the user account",
        "tags": [
          "Manager"
        ]
      }
    },
    "/user/new_employees": {
      "get": {
        "description": "This endpoint requires a minimun security clearance of level 5.",
        "operationId": "UserController_viewNewEmployees",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Here are the list of new employees without user accounts"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get the new employees without user accounts",
        "tags": [
          "Manager"
        ]
      }
    },
...
  },
  "info": {
    "title": "ABAS v3 API",
    "description": "API documentation for ABAS v3 project",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      },
      "security_clearance": {
        "type": "apiKey",
        "name": "Security-Clearance",
        "in": "header",
        "description": "Required security clearance level (documentational only)."
      }
    },
    "schemas": {
...
      "UserDetailsDto": {
        "type": "object",
        "properties": {
          "employee_id": {
            "type": "string",
            "example": "ABISC-250710-001",
            "description": "The ID of the employee to link this user to"
          },
          "username": {
            "type": "string",
            "example": "sample_username",
            "description": "The username for the new user"
          },
          "email": {
            "type": "string",
            "example": "sample@gmail.com",
            "description": "Email address for the new user"
          },
          "password": {
            "type": "string",
            "example": "avegabros",
            "description": "Password for the new user"
          }
        },
        "required": [
          "employee_id",
          "username",
          "email",
          "password"
        ]
      },
      "CreateUserWithRolePermissionDto": {
        "type": "object",
        "properties": {
          "user_details": {
            "$ref": "#/components/schemas/UserDetailsDto"
          },
          "role_permission_ids": {
            "example": [
              1
            ],
            "description": "The role permission IDs to assign to the user",
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "user_details",
          "role_permission_ids"
        ]
      },
      "AddUserRolePermissionsDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "example": 1,
            "description": "The id of the user to add role permission"
          },
          "rolePermissionIds": {
            "example": [
              101,
              102,
              103
            ],
            "description": "The role permission id that will be assign or added to the user it is array because you can add multiple role permission",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "userId",
          "rolePermissionIds"
        ]
      },
      "UserEmailResetTokenDto": {
        "type": "object",
        "properties": {}
      },
      "DeactivateUserAccountDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "number",
            "example": 1,
            "description": "The id number of the user you want to deactivate"
          }
        },
        "required": [
          "user_id"
        ]
      },
      "ReactivateUserAccountDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "number",
            "example": 1,
            "description": "The id number of the user you want to reactivate"
          }
        },
        "required": [
          "user_id"
        ]
      },
...
    }
  }
}